Add a cast to avoid a warning and potential bug and/or crash.
This is a temporary, minimal change; will solve this properly
in a follow-up commit.

* subversion/libsvn_fs_fs/verify.c (compare_p2l_to_l2p_index):
   Cast entry->item.number to long to match the printf format,
   like other similar cases in this file.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1640966 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs_fs/verify.c
@@ -405,7 +405,7 @@ compare_p2l_to_l2p_index(svn_fs_t *fs,
                                            "refers to revision %ld outside "
                                            "the rev / pack file (%ld-%ld)"),
                                          apr_off_t_toa(pool, entry->offset),
-                                         entry->item.number,
+                                         (long)entry->item.number,
                                          start, start + count - 1);
             }
           else
